+++ core/src/main/java/uk/gov/justice/services/core/featurecontrol/DisabledFeatureException.java
@@ -1,0 +1,10 @@
+package uk.gov.justice.services.core.featurecontrol;
+
+import uk.gov.justice.services.common.exception.ForbiddenRequestException;
+
+public class DisabledFeatureException extends ForbiddenRequestException {
+
+    public DisabledFeatureException(final String message) {
+        super(message);
+    }
+}