+++ core/src/test/java/uk/gov/justice/services/core/handler/registry/TestHandlerRegistryCacheProducer.java
@@ -1,0 +1,18 @@
+package uk.gov.justice.services.core.handler.registry;
+
+import uk.gov.justice.services.core.featurecontrol.FeatureControlAnnotationFinder;
+import uk.gov.justice.services.core.handler.registry.HandlerRegistryCache;
+
+import javax.enterprise.inject.Produces;
+
+/**
+ * Test version of HandlerRegistryCacheProducer to stop open ejb freaking out about
+ * @Singletons in its web scope
+ */
+public class TestHandlerRegistryCacheProducer {
+
+    @Produces
+    public HandlerRegistryCache handlerRegistryCache() {
+        return new HandlerRegistryCache(new FeatureControlAnnotationFinder());
+    }
+}